Fixes integer overfloww error when using large tick sizes

+++ Project 1/src/consumer/consumer.go
@@ -7,12 +7,12 @@ import (
 
 var logger *log.Logger
 var (
-	next_tick      int = 1
-	wait_time_tick int // in milliseconds
+	next_tick      float64 = 1
+	wait_time_tick float64 // in milliseconds
 	queue          stats.QueueMgr
 )
 
-func Init(l *log.Logger, wait_tick int) *stats.QueueMgr {
+func Init(l *log.Logger, wait_tick float64) *stats.QueueMgr {
 	logger = l
 	logger.Println("[Consumer] Started")
 
@@ -23,13 +23,13 @@ func Init(l *log.Logger, wait_tick int) *stats.QueueMgr {
 	return &queue
 }
 
-func Tick(t int) {
+func Tick(t float64) {
 	if next_tick <= t {
 		consume_packet(t)
 	}
 }
 
-func consume_packet(t int) {
+func consume_packet(t float64) {
 	//logger.Println("[Consumer] Consuming Packet for", wait_time_tick, "ticks")
 	if queue.Size > 0 {
 		var p stats.Packet
